AD/Barry: This says “recommended”, and not even “RECOMMENDED”, but later we have, “If the ZONEMD RRSet contains more than one RR with the same Scheme and Hash Algorithm, digest verification MUST NOT be considered successful.”  So how is this not a MUST, given that it will not interoperate if it’s violated?
Authors: The difference is that the later text is specifically referring
to RRs that have the same scheme and algorithm, but the earlier one
is not.

AD/Barry: I think the text should be clarified a little, to avoid what
appears to be a conflict.  It should just take a few words, no?

Authors: Words added (repeated)