AD/Barry: It’s not wrong, but it’s slightly jarring that all the items around this say “MUST” and this one says “SHALL”.  Any reason, or should we switch this to “MUST” to match the others?
Authors: Yes, the wording here is tricky.  Originally it was written like
<this> But that wasn't clear that you should include one instance of the
duplicated RR.  It would probably be more clear in active voice such as
"... MUST include only one instance of duplicate RRs..." but all these
bullets are passive voice.  Open to suggestions.

AD/Barry: Hm, I see.  All I had been suggesting was changing "SHALL" to "MUST"
in order to be consistent, but now I see that it's a little more
complicated.  On first thinking, it seems that the original (with the
"MUST NOT") is clearer.  How about this?: "If there are duplicate RRs
with equal owner, class, type, and RDATA, only one instance is included,
and the duplicates MUST be omitted."

Authors: Sounds good.